apps/browser-agent/pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
description = "An agent for interacting with a browser so you don't have to"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.1.23",
    "langchain>=0.3.14",
    "langchain-openai>=0.2.14",
    "playwright>=1.49.1",
    "portkey-ai>=1.10.1",
    "python-dotenv>=1.0.1",
]
